@@ -4,16 +4,18 @@
 	"description": "Find a file by walking up parent directories",
 	"license": "MIT",
 	"repository": "sindresorhus/find-up-cli",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
 	"bin": {
-		"find-up": "cli.js"
+		"find-up": "./cli.js"
 	},
 	"engines": {
-		"node": ">=8"
+		"node": "^12.20.0 || ^14.13.1 || >=16.0.0"
 	},
 	"scripts": {
 		"test": "xo && ava"
@@ -44,12 +46,12 @@
 		"path"
 	],
 	"dependencies": {
-		"find-up": "^4.0.0",
-		"meow": "^5.0.0"
+		"find-up": "^6.0.0",
+		"meow": "^10.1.1"
 	},
 	"devDependencies": {
-		"ava": "^1.4.1",
-		"execa": "^1.0.0",
-		"xo": "^0.24.0"
+		"ava": "^3.15.0",
+		"execa": "^5.1.1",
+		"xo": "^0.44.0"
 	}
 }
